+++ lib/solid_lints.dart
@@ -1,6 +1,7 @@
 library solid_metrics;
 
 import 'package:custom_lint_builder/custom_lint_builder.dart';
+import 'package:solid_lints/src/lints/avoid_debug_print/avoid_debug_print_rule.dart';
 import 'package:solid_lints/src/lints/avoid_global_state/avoid_global_state_rule.dart';
 import 'package:solid_lints/src/lints/avoid_late_keyword/avoid_late_keyword_rule.dart';
 import 'package:solid_lints/src/lints/avoid_non_null_assertion/avoid_non_null_assertion_rule.dart';
@@ -59,6 +60,7 @@ class _SolidLints extends PluginBase {
       PreferLastRule.createRule(configs),
       PreferMatchFileNameRule.createRule(configs),
       ProperSuperCallsRule.createRule(configs),
+      AvoidDebugPrint.createRule(configs),
     ];
 
     // Return only enabled rules