+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Added `avoid_using_debug_print` rule
 - Fixed an issue with no_magic_number lint
-- Fixed `avoid_unused_parameters` lint's bug regarding function variable assignment/declaration 
+- Fixed `avoid_unused_parameters` to report positional parameters from typedef if their name are not underscores.
 
 ## 0.1.4