+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Added `avoid_debug_print` rule
 - Fixed an issue with no_magic_number lint
+- Improvement for `avoid_returning_widget` lint:
+  - ignores methods that override ones that return widget (build() for example)
+  - no longer allows returning widgets from methods/functions named build 
 
 ## 0.1.4