+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## 0.1.5
 
-- Added `avoid_using_debug_print` rule
+- Added `avoid_debug_print` rule
 - Fixed an issue with no_magic_number lint
 - Fixed `avoid_unused_parameters` to report positional parameters from typedef if their name are not underscores.
+- Improvement for `avoid_returning_widget` lint:
+  - ignores methods that override ones that return widget (build() for example)
+  - no longer allows returning widgets from methods/functions named build 
 
 ## 0.1.4