+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   - ignores methods that override ones that return widget (build() for example)
   - no longer allows returning widgets from methods/functions named build
 - Fixed unexpected avoid_unnecessary_type_assertions
+- Added `excludeNames` param for `function_lines_of_code` lint
 
 ## 0.1.4